+++ src/main/java/org/gxfj/iknow/pojo/Comment.java
@@ -6,8 +6,8 @@ import java.util.Date;
 
 @Entity
 public class Comment {
-    public static final int COMMENT_UN_DELETED = 0;
-    public static final int COMMENT_DELETED = 1;
+    public static final byte COMMENT_UN_DELETED = 0;
+    public static final byte COMMENT_DELETED = 1;
 
     private Integer id;
     private String content;
